Update: fastPath for fill
